move those codes of about DataFlowInfo to ZookeeperMetaWatcher.
